poetry_corpus/scripts/bow.py
import copy
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer


def bow(train_texts, test_texts, use_tfidf=False, max_features=None, bow_ngrams=(1,2), analyzer='word'):
    train = copy.deepcopy(train_texts)
    test = copy.deepcopy(test_texts)

    if use_tfidf:
        vectorizer = TfidfVectorizer(analyzer=analyzer, ngram_range=bow_ngrams, max_features=max_features)
    else:
        vectorizer = CountVectorizer(analyzer=analyzer, ngram_range=bow_ngrams, max_features=max_features)
    data = train+test
    data = vectorizer.fit_transform(data)
    train_data = data[:len(train)]
    test_data = data[len(train):]
    return train_data, test_data