requirements.txt
lxml
cryptography
scrapy >= 1.2
dicttoxml
Django
numpy
scipy
scikit-learn
django-reversion